+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [0.0.40] - 2021-03-01
+### 🚀 Features
+- Add WaitSpot resources ([#300](https://github.com/whirli/whirli-client/pull/300))
+- Add gtin to product interface ([#299](https://github.com/whirli/whirli-client/pull/299))
 ## [0.0.39] - 2020-12-07
 ### 🏡 Chore
 - Update Subscription Pricing Plan properties ([#296](https://github.com/whirli/whirli-client/pull/296))